+++ Stackage/PerformBuild.hs
@@ -637,8 +637,7 @@ singleBuild pb@PerformBuild {..} registeredPackages SingleBuild {..} = do
                                     ]
                         else do
                             outH <- getOutH
-                            hPutStrLn outH $ "Test suite not built: " ++ test
-                            hFlush outH
+                            hPut outH $ encodeUtf8 $ "Test suite not built: " ++ pack test
 
             savePreviousResult pb Test pident $ either (const False) (const True) eres
             case (eres, pcTests) of