+++ ChangeLog.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.15.0.0
 
 * Ignore packages with invalid flag definitions [fpco/stackage#2468](https://github.com/fpco/stackage/issues/2468)
 * Support a `no-revisions` field to ignore Hackage revisions (experimental) [fpco/stackage#2469](https://github.com/fpco/stackage/issues/2469)